requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
numpy==2.0.0
pillow==10.4.0
pypresence==4.3.0
requests==2.28.2
sv-ttk==2.6.0
urllib3==1.26.19
zstandard==0.23.0